requirements.txt
beautifulsoup4==4.6.0
python-wordpress-xmlrpc